# backend/openapi.yaml
openapi: 3.1.0
info:
  title: RAG Assistant Backend
  version: 0.2.1
paths:
...
  /api/stores:
    get:
      tags:
      - stores
      summary: List Stores
      operationId: list_stores_api_stores_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StoreOut'
                title: Response List Stores Api Stores Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - stores
      summary: Create Store
      operationId: create_store_api_stores_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/stores/{store_id}:
    delete:
      tags:
      - stores
      summary: Delete Store
      operationId: delete_store_api_stores__store_id__delete
      parameters:
      - name: store_id
        in: path
        required: true
        schema:
          type: integer
          title: Store Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/stores/{store_id}/restore:
    post:
      tags:
      - stores
      summary: Restore Store
      operationId: restore_store_api_stores__store_id__restore_post
      parameters:
      - name: store_id
        in: path
        required: true
        schema:
          type: integer
          title: Store Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/documents/store/{store_id}:
    get:
      tags:
      - documents
      summary: List Documents For Store
      operationId: list_documents_for_store_api_documents_store__store_id__get
      parameters:
      - name: store_id
        in: path
        required: true
        schema:
          type: integer
          title: Store Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentOut'
                title: Response List Documents For Store Api Documents Store  Store
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/documents/{document_id}:
    delete:
      tags:
      - documents
      summary: Delete Document
      operationId: delete_document_api_documents__document_id__delete
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: integer
          title: Document Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/documents/{document_id}/restore:
    post:
      tags:
      - documents
      summary: Restore Document
      operationId: restore_document_api_documents__document_id__restore_post
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: integer
          title: Document Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DocumentOut:
      properties:
        id:
          type: integer
          title: Id
        store_id:
          type: integer
          title: Store Id
        filename:
          type: string
          maxLength: 255
          title: Filename
        display_name:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Display Name
        status:
          $ref: '#/components/schemas/DocumentStatus'
        size_bytes:
          type: integer
          title: Size Bytes
        created_at:
          type: string
          format: date-time
          title: Created At
        gcs_uri:
          anyOf:
          - type: string
            maxLength: 512
          - type: 'null'
          title: Gcs Uri
      type: object
      required:
      - id
      - store_id
      - filename
      - status
      - size_bytes
      - created_at
      title: DocumentOut
    DocumentStatus:
      type: string
      enum:
      - PENDING
      - RUNNING
      - DONE
      - ERROR
      title: DocumentStatus
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StoreCreate:
      properties:
        display_name:
          type: string
          maxLength: 100
          minLength: 1
          title: Display Name
      type: object
      required:
      - display_name
      title: StoreCreate
    StoreOut:
      properties:
        id:
          type: integer
          title: Id
        display_name:
          type: string
          title: Display Name
        fs_name:
          type: string
          title: Fs Name
      type: object
      required:
      - id
      - display_name
      - fs_name
      title: StoreOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...